+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
 ### Latest changes
 
+* VFS-7724 Changed separate dataset and archives views into splitted browser
 * VFS-8698 Fixed ESLint warnings
 * VFS-8755 Fixed randomly failing tests
 * VFS-8639 Removed iterator strategy from workflows